move try_extend into its own module and add test
